Revert "Switch between the list/grid preview"
