+++ Backend/src/main/java/com/developer/auctionapp/configuration/SecurityConfig.java
@@ -59,7 +59,6 @@ public class SecurityConfig {
                 .antMatchers("/auctionapp/image/getAll").permitAll()
                 .antMatchers("/auctionapp/subcategory/getAll").permitAll()
                 .antMatchers("/auctionapp/product/getProductFromId").permitAll()
-                .antMatchers("/auctionapp/bid/getAll").permitAll()
                 .anyRequest().authenticated()
                 .and()
                 .httpBasic();
