+++ Backend/src/main/java/com/developer/auctionapp/entity/User.java
@@ -106,6 +106,7 @@ public class User {
         this.gender = gender;
         this.dateOfBirth = dateOfBirth;
         address = null;
+
     }
 
     public User(
