+++ Backend/src/main/java/com/developer/auctionapp/service/InitializeService.java
@@ -10,6 +10,7 @@ public interface InitializeService {
 
     Boolean initializeCategoryTable();
 
+
     void initializeSubcategoryTable();
 
     void initializeRoleTable();
