+++ Backend/src/main/java/com/developer/auctionapp/service/impl/InitializeServiceImpl.java
@@ -42,6 +42,7 @@ public class InitializeServiceImpl implements InitializeService {
 
     private String initializeRole = "";
 
+
     /**
      * A method that inserts data into the Category table after checking that no data already exists in the table
      */
@@ -116,7 +117,6 @@ public class InitializeServiceImpl implements InitializeService {
     }
 
     /**
-
      * A method that inserts data into the User and Role tables after checking that no data already exists in the table
      */
