+++ Backend/src/main/java/com/developer/auctionapp/service/impl/ProductServiceImpl.java
@@ -40,6 +40,8 @@ public class ProductServiceImpl implements ProductService {
 
     private final BidRepository bidRepository;
 
+    private final UserRepository userRepository;
+
     public ProductServiceImpl(
             final ProductRepository productRepository,
             final ImageRepository imageRepository,
@@ -53,6 +55,7 @@ public class ProductServiceImpl implements ProductService {
         this.subcategoryRepository = subcategoryRepository;
         this.userService = userService;
         this.bidRepository = bidRepository;
+        this.userRepository = userRepository;
     }
 
     /**
