+++ Backend/src/main/java/com/developer/auctionapp/service/impl/UserServiceImpl.java
@@ -23,6 +23,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
