+++ Frontend/src/app/app.component.html
@@ -16,9 +16,9 @@
     *ngIf="!hasRoute('Product') && !hasRoute('Login') && !hasRoute('Register') && !hasRoute('Category') && !hasRoute('myAccount')  && !hasRoute('Product')">
     <app-middle></app-middle>
   </div>
-  <app-tab [listOfProductsImages]="listOfProductsImages"
+  <app-tab
     *ngIf="!hasRoute('Login') && !hasRoute('Register') &&  !hasRoute('Category')  && !hasRoute('myAccount')  && !hasRoute('Product')"></app-tab>
-  <app-recommended-tab [listOfProductsImages]="listOfProductsImages"
+  <app-recommended-tab
     *ngIf="authG.canActivate()  && !hasRoute('myAccount')  && !hasRoute('Product')"></app-recommended-tab>
   <app-footer></app-footer>
 </body>
