+++ Frontend/src/app/app.module.ts
@@ -44,64 +44,62 @@ import { BidderItemComponent } from './core/components/bidder-item/bidder-item.c
 import {MatSelectModule} from '@angular/material/select';
 import {MatFormFieldModule} from '@angular/material/form-field'
 import {MatAutocompleteModule} from '@angular/material/autocomplete';
-import { NgxPaginationModule } from 'ngx-pagination';
-import { ItemListComponent } from "./core/components/item-list/item-list.component";
+import { NgxPaginationModule } from 'ngx-pagination'; 
 
 @NgModule({
-    declarations: [
-        AppComponent,
-        TabComponent,
-        MiddleComponent,
-        DetailsComponent,
-        ItemComponent,
-        LastChanceComponent,
-        NewArrivalsComponent,
-        SidemenuComponent,
-        NavbarComponent,
-        FooterComponent,
-        SubcategoryComponent,
-        ProductOverviewComponent,
-        ImagesComponent,
-        NotificationComponent,
-        ShopComponent,
-        RecommendedProductsComponent,
-        RecommendedTabComponent,
-        RegisterFormComponent,
-        LoginFormComponent,
-        MyAccountComponent,
-        SettingsComponent,
-        SellerTabComponent,
-        BidTabComponent,
-        ProfileComponent,
-        TableItemComponent,
-        SocialNetworkLogComponent,
-        SocialNetworkBtnsComponent,
-        DatePickerBirthdayComponent,
-        AddressCardInformationComponent,
-        AddItemComponent,
-        AddItemComponent,
-        ShippingComponent,
-        BasicComponent,
-        AddItemComponent,
-        BidderItemComponent,
-        ItemListComponent
-    ],
-    providers: [],
-    bootstrap: [AppComponent],
-    imports: [
-        BrowserModule,
-        AppRoutingModule,
-        BrowserAnimationsModule,
-        MatSidenavModule,
-        MatTabsModule,
-        MatMenuModule,
-        HttpClientModule,
-        FormsModule,
-        MatButtonModule,
-        MatFormFieldModule,
-        MatSelectModule,
-        MatAutocompleteModule,
-        NgxPaginationModule
-    ]
+  declarations: [
+    AppComponent,
+    TabComponent,
+    MiddleComponent,
+    DetailsComponent,
+    ItemComponent,
+    LastChanceComponent,
+    NewArrivalsComponent,
+    SidemenuComponent,
+    NavbarComponent,
+    FooterComponent,
+    SubcategoryComponent,
+    ProductOverviewComponent,
+    ImagesComponent,
+    NotificationComponent,
+    ShopComponent,
+    RecommendedProductsComponent,
+    RecommendedTabComponent,
+    RegisterFormComponent,
+    LoginFormComponent,
+    MyAccountComponent,
+    SettingsComponent,
+    SellerTabComponent,
+    BidTabComponent,
+    ProfileComponent,
+    TableItemComponent,
+    SocialNetworkLogComponent,
+    SocialNetworkBtnsComponent,
+    DatePickerBirthdayComponent,
+    AddressCardInformationComponent,
+    AddItemComponent,
+    AddItemComponent,
+    ShippingComponent,
+    BasicComponent,
+    AddItemComponent,
+    BidderItemComponent
+  ],
+  imports: [
+    BrowserModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    MatSidenavModule,
+    MatTabsModule,
+    MatMenuModule,
+    HttpClientModule,
+    FormsModule,
+    MatButtonModule,
+    MatFormFieldModule,
+    MatSelectModule,
+    MatAutocompleteModule,
+    NgxPaginationModule
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
 })
 export class AppModule { }
