+++ Frontend/src/app/core/components/add-item/add-item.component.ts
@@ -9,7 +9,6 @@ import { ShippingComponent } from '../shipping/shipping.component';
   templateUrl: './add-item.component.html',
   styleUrls: ['./add-item.component.css']
 })
-
 export class AddItemComponent implements OnInit {
 
   @ViewChild('container', { read: ViewContainerRef })
