+++ Frontend/src/app/core/components/item/item.component.ts
@@ -18,6 +18,14 @@ export class ItemComponent implements OnInit {
   constructor(private router: Router) { }
 
   ngOnInit(): void {
+    const storage = getStorage();
+    for (const img of this.product.imageName) {
+      getDownloadURL(ref(storage, img))
+        .then((url) => {
+          this.image.push(url);
+        })
+    }
+    this.product.imageName = this.image;
   }
 
   onClick(product1: Product): void {
