+++ Frontend/src/app/core/components/last-chance/last-chance.component.ts
@@ -1,9 +1,7 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Product } from '../../models/product';
-import { ProductImages } from '../../models/product-images';
 import { ApiService } from '../../services/api.service';
 import { BidService } from '../../services/bid.service';
-import { ProductUtils } from '../../utils/product-utils';
 
 @Component({
   selector: 'app-last-chance',
@@ -17,9 +15,6 @@ export class LastChanceComponent implements OnInit {
 
   usersRole: number = 0;
 
-  @Input()
-  listOfProductsImages: ProductImages[] = [];
-
   constructor(private bidService: BidService,
     private apiService: ApiService) { }
 
@@ -27,10 +22,8 @@ export class LastChanceComponent implements OnInit {
     this.apiService.getLastChanceProduct().subscribe((rez) => {
       let products = <Product[]>JSON.parse(JSON.stringify(rez));
       this.lastChanceProducts = products.filter(item => item.status.toString() == 'false');
-      setTimeout(() => {
-        this.lastChanceProducts = ProductUtils.productsWithListOfImages(this.lastChanceProducts,this.listOfProductsImages)
-      }, 1000);
     })
     this.bidService.getBids();
   }
 }
+
