+++ Frontend/src/app/core/components/product-overview/product-overview.component.html
@@ -49,6 +49,7 @@
             </div>
             <div class="bidInfo" *ngIf="displayPaymentButton == true">
                 <button (click)="makePayment(highestBid)" class="button">PAY</button>
+
             </div>
             <div class="bidInfo" *ngIf="displayPaymentButton == true">
                 <button (click)="makePayment(highestBid)" class="button">PAY</button>
