+++ Frontend/src/app/core/components/productoverview/productoverview.component.ts
@@ -177,7 +177,7 @@ export class ProductOverviewComponent implements OnInit {
 
 
   /**
-   * The method we use it to create payment in Stripe
+   * The method we use it to create payment in Stripe 
    * @param amount we are paying for the product
    * @returns in case that the user has already paid for this product
    */
