+++ Frontend/src/app/core/components/recommended-products/recommended-products.component.ts
@@ -1,11 +1,9 @@
 import { JsonPipe } from '@angular/common';
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Bid } from '../../models/bid';
 import { Product } from '../../models/product';
-import { ProductImages } from '../../models/product-images';
 import { User } from '../../models/user';
 import { ApiService } from '../../services/api.service';
-import { ProductUtils } from '../../utils/product-utils';
 
 @Component({
   selector: 'app-recommended-products',
@@ -28,29 +26,32 @@ export class RecommendedProductsComponent implements OnInit {
 
   recommendedProducts: Product[] = []
 
-  @Input()
-  listOfProductsImages: ProductImages[] = [];
-
   constructor(private apiService: ApiService) { }
 
   ngOnInit(): void {
     if (localStorage.getItem('token') != null) {
       this.apiService.getCurrentUser().subscribe((user) => {
         this.user = <User>JSON.parse(JSON.stringify(user));
-        this.apiService.getAllBids().subscribe((bidResponse) => {
-          this.bids = <Bid[]>JSON.parse(JSON.stringify(bidResponse));
-          this.apiService.getAllProducts().subscribe((productResponse) => {
-            this.products = <Product[]>JSON.parse(JSON.stringify(productResponse));
-            setTimeout(() => {
-              this.products = ProductUtils.productsWithListOfImages(this.products, this.listOfProductsImages)
-            }, 1000);
-            this.findSimilarProductsFromSelling(this.products);
-            this.findSimilarProductsFromBidding(this.bids, this.products);
-            this.recommendedProducts = this.findRecommendedProducts();
-          })
-        })
       })
     }
+    this.apiService.getAllBids().subscribe((bidResponse) => {
+      this.bids = <Bid[]>JSON.parse(JSON.stringify(bidResponse));
+      this.apiService.getAllProducts().subscribe((productResponse) => {
+        this.products = <Product[]>JSON.parse(JSON.stringify(productResponse));
+        this.findSimilarProductsFromSelling(this.products);
+        this.findSimilarProductsFromBidding(this.bids, this.products);
+        this.recommendedProducts = this.products.filter((item) => item.userId != this.user.id &&
+          (this.listOfCategories.includes(item.categoryId) || this.listOfSubcategories.includes(item.subcategoryId)))
+        if (this.recommendedProducts.length < 8) {
+          for (var i = 0; i <= 8; i++) {
+            this.recommendedProducts.push(this.products[i]);
+            if (this.recommendedProducts.length == 8) {
+              break;
+            }
+          }
+        }
+      })
+    })
   }
 
   /**
@@ -85,24 +86,5 @@ export class RecommendedProductsComponent implements OnInit {
         this.listOfSubcategories.push(item.subcategoryId);
       }
     })
-  }
-
-  /**
-   * A method that finds recommended products for specific user
-   * @returns list of products
-   */
-
-  findRecommendedProducts() {
-    let recProducts = this.products.filter((item) => item.userId != this.user.id &&
-      (this.listOfCategories.includes(item.categoryId) || this.listOfSubcategories.includes(item.subcategoryId)))
-    if (recProducts.length < 8) {
-      for (var i = 0; i <= 8; i++) {
-        recProducts.push(this.products[i]);
-        if (recProducts.length == 8) {
-          break;
-        }
-      }
-    }
-    return recProducts
   }
 }
