+++ Frontend/src/app/core/components/recommended-tab/recommended-tab.component.html
@@ -10,7 +10,7 @@
         <mat-tab-group>
             <mat-tab label="Recommended Products">
                 <div class="firstTab">
-                    <app-recommended-products [listOfProductsImages]="listOfProductsImages"></app-recommended-products>
+                    <app-recommended-products></app-recommended-products>
                 </div>
             </mat-tab>
         </mat-tab-group>
