+++ Frontend/src/app/core/components/tab/tab.component.html
@@ -10,12 +10,12 @@
         <mat-tab-group>
             <mat-tab label="New Arrivals">
                 <div class="firstTab">
-                    <app-new-arrivals [listOfProductsImages]="listOfProductsImages"></app-new-arrivals>
+                    <app-new-arrivals></app-new-arrivals>
                 </div>
             </mat-tab>
             <mat-tab label="Last Chance">
                 <div class="secondTab">
-                    <app-last-chance [listOfProductsImages]="listOfProductsImages"></app-last-chance>
+                    <app-last-chance></app-last-chance>
                 </div>
             </mat-tab>
         </mat-tab-group>
