--- Frontend/src/app/core/models/product-images.spec.ts
@@ -1,7 +1,0 @@
-import { ProductImages } from './product-images';
-
-describe('ProductImages', () => {
-  it('should create an instance', () => {
-    expect(new ProductImages()).toBeTruthy();
-  });
-});
